--- src/main/avro/selection.avsc
@@ -1,18 +1,0 @@
-{
-  "imports": [
-    "references/platform_identical_to.avsc"
-  ],
-  "name": "Selection",
-  "namespace": "com.maria.messaging",
-  "type": "record",
-  "fields": [
-    {
-      "name": "id",
-      "type": "string"
-    },
-    {
-      "name": "name",
-      "type": "string"
-    }
-  ]
-}
