+++ src/main/resources/application.properties
@@ -1,5 +1,4 @@
 server.port=8082
-#spring.kafka.streams.properties.schema.registry.url = https://localhost:8083
 spring.h2.console.enabled=true
 spring.sql.init.platform=postgres
 spring.datasource.url=jdbc:postgresql://localhost:5432/postgres
@@ -13,5 +12,4 @@ spring.data.web.pageable.default-page-size=3
 spring.jpa.show-sql=true
 spring.kafka.bootstrapServers=localhost:9092
 key.serializer=org.apache.kafka.common.serialization.StringSerializer
-#value.serializer=org.apache.kafka.common.serialization.StringSerializer
-value.serializer= io.confluent.kafka.serializers.KafkaAvroSerializer
+value.serializer=io.confluent.kafka.serializers.KafkaAvroSerializer
