+++ src/test/java/com/maria/crudapp_participants/service/EventServiceImplTest.java
@@ -4,8 +4,11 @@ import com.maria.crudapp_participants.dto.ShortEvent;
 import com.maria.crudapp_participants.entity.Event;
 import com.maria.crudapp_participants.entity.Market;
 import com.maria.crudapp_participants.entity.Participant;
+import com.maria.crudapp_participants.facade.ParticipantFacadeImpl;
 import com.maria.crudapp_participants.repository.EventRepository;
 import com.maria.crudapp_participants.selections.Selection;
+import com.maria.crudapp_participants.service.EventServiceImpl;
+import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -14,6 +17,7 @@ import org.junit.jupiter.params.provider.MethodSource;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.core.convert.converter.Converter;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.PageRequest;
@@ -21,15 +25,11 @@ import org.springframework.data.domain.Pageable;
 
 import java.math.BigDecimal;
 import java.sql.Date;
-import java.util.Arrays;
-import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
+import java.util.*;
 import java.util.stream.Stream;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
+import static org.mockito.Mockito.*;
 
 
 @ExtendWith(MockitoExtension.class)
@@ -318,5 +318,92 @@ class EventServiceImplTest {
         when(eventRepository.findById(UUID.fromString("f15ef034-4f9a-4be1-a0c0-dd09e154e48d"))).thenReturn(Optional.of(generateEvent()));
         BigDecimal actual = eventService.maxPayoutPerEvent(UUID.fromString("f15ef034-4f9a-4be1-a0c0-dd09e154e48d"));
         assertThat(actual).isEqualTo(BigDecimal.valueOf(10.5));
+    }
+
+    @ExtendWith(MockitoExtension.class)
+    static
+    class ParticipantFacadeImplTest {
+
+        private final ParticipantService participantService;
+        private final Converter<Participant, ParticipantDTO> participantToParticipantDTOConverter;
+        private final Converter<ParticipantDTO, Participant> participantDTOToParticipantConverter;
+        private final ParticipantFacadeImpl participantFacade;
+
+        public ParticipantFacadeImplTest() {
+            this.participantService = mock(ParticipantService.class);
+            this.participantToParticipantDTOConverter = mock(Converter.class);
+            this.participantDTOToParticipantConverter = mock(Converter.class);
+            this.participantFacade = new ParticipantFacadeImpl(participantService, participantToParticipantDTOConverter, participantDTOToParticipantConverter);
+        }
+
+        @Test
+        void saveParticipant() {
+            ParticipantDTO participantDto = mock(ParticipantDTO.class);
+            Participant participant = mock(Participant.class);
+
+            when(participantDTOToParticipantConverter.convert(participantDto)).thenReturn(participant);
+            when(participantService.saveParticipant(participant)).thenReturn(participant);
+            when(participantToParticipantDTOConverter.convert(participant)).thenReturn(participantDto);
+
+            ParticipantDTO participantDTO = participantFacade.saveParticipant(participantDto);
+
+            assertThat(participantDTO).isEqualTo(participantDto);
+            verify(participantService, times(1)).saveParticipant(participant);
+        }
+
+        @Test
+        void fetchParticipantsList() {
+            Pageable pageable = PageRequest.of(0, 3);
+            String searchString = "Sh";
+            Page<Participant> participants = mock(Page.class);
+            List<ParticipantDTO> participantDto = new ArrayList<>();
+
+
+            when(participantService.fetchParticipantsList(searchString, pageable)).thenReturn(participants);
+            when(participants.getContent().stream().map(participantToParticipantDTOConverter::convert).toList()).thenReturn(participantDto);
+            when(participants.getPageable()).thenReturn(pageable);
+            when(participants.getTotalElements()).thenReturn(3L);
+
+            Page<ParticipantDTO> expected = new PageImpl<>(participantDto, pageable, 3L);
+            Page<ParticipantDTO> actual = participantFacade.fetchParticipantsList(searchString, pageable);
+
+            Assertions.assertThat(actual).isEqualTo(expected);
+            verify(participantService, times(1)).fetchParticipantsList(searchString, pageable);
+        }
+
+        @Test
+        void updateParticipant() {
+            ParticipantDTO participantDto = mock(ParticipantDTO.class);
+            Participant participant = mock(Participant.class);
+
+            when(participantDTOToParticipantConverter.convert(participantDto)).thenReturn(participant);
+            when(participantService.updateParticipant(participant, participant.getId())).thenReturn(participant);
+            when(participantToParticipantDTOConverter.convert(participant)).thenReturn(participantDto);
+
+            ParticipantDTO participantDTO = participantFacade.updateParticipant(participantDto, participantDto.getId());
+
+            assertThat(participantDTO).isEqualTo(participantDto);
+            verify(participantService, times(1)).updateParticipant(participant, participant.getId());
+        }
+
+        @Test
+        void deleteParticipantById() {
+            participantFacade.deleteParticipantById(UUID.fromString("9d9239ac-1257-11ed-861d-0242ac120002"));
+            verify(participantService, times(1)).deleteParticipantById(UUID.fromString("9d9239ac-1257-11ed-861d-0242ac120002"));
+        }
+
+        @Test
+        void findParticipantById() {
+            ParticipantDTO participantDto = mock(ParticipantDTO.class);
+            Participant participant = mock(Participant.class);
+
+            when(participantService.findParticipantById(UUID.fromString("9d9239ac-1257-11ed-861d-0242ac120002"))).thenReturn(participant);
+            when(participantToParticipantDTOConverter.convert(participant)).thenReturn(participantDto);
+
+            ParticipantDTO participantDTO = participantFacade.findParticipantById(UUID.fromString("9d9239ac-1257-11ed-861d-0242ac120002"));
+
+            assertThat(participantDTO).isEqualTo(participantDto);
+            verify(participantService, times(1)).findParticipantById(UUID.fromString("9d9239ac-1257-11ed-861d-0242ac120002"));
+        }
     }
 }
